+++ src/app/mood/page.tsx
@@ -1,0 +1,3 @@
+export default function MoodPicker() {
+  return <MoodPicker />;
+}
